requirements/base.txt
Django>=2.2
djangorestframework>=3.9

psycopg2-binary

django-extensions
